#70 - Fix all errors related to deleting linked objects - This patches all but 9 of the tests, mostly related to dealing with the linked objects. - Also patches deleting students with us removing all links BEFORE going for the delete. - Remaining testing issues are related to the links and duplicate keys related to the linked objects (this needs more research into why this is occuring) - Also implements the migration to pure GUIDs across everything, including students (Need to fix the add fields still but everything in teh backend has been changed. Thought this might have been an issue with some of the strange Id stuff)

+++ RecruiterPathway/Repository/RecruiterRepository.cs
@@ -33,7 +33,7 @@ namespace RecruiterPathway.Repository
         {
             var recruiter = await GetById(id);
             //Delete all the linked objects
-            lock (_context)
+            lock (_context) lock(_userManager)
             {
                 var comments = _context.Comment.Where(c => c.Recruiter.Id == recruiter.Id);
                 _context.Comment.RemoveRange(comments);
@@ -42,10 +42,9 @@ namespace RecruiterPathway.Repository
                 var watches = _context.WatchList.Where(c => c.Recruiter.Id == recruiter.Id);
                 _context.WatchList.RemoveRange(watches);
                 _context.SaveChanges();
-                //Delete Student MANUALLY to prevent issues with GetById
-                _context.Recruiter.Remove(recruiter);
-                _context.SaveChanges();
             }
+            //Manual Recruiter delete through userManager to resolve all link issues
+            await _userManager.DeleteAsync(recruiter);
         }
 
         override public async ValueTask<Recruiter> GetById(object id)
@@ -108,7 +107,7 @@ namespace RecruiterPathway.Repository
             }
             return recruiter;
         }
-
+        //This method can only be used in Prod since ClaimsPrincipal is not really mockable
         override async public Task<Recruiter> GetSignedInRecruiter(System.Security.Claims.ClaimsPrincipal principal, bool getExtras)
         {
             var recruiterBasedOnSignin = await _userManager.GetUserAsync(principal);
@@ -145,17 +144,37 @@ namespace RecruiterPathway.Repository
             {
                 return false;
             }
-            var pstatus = recruiter.PipelineStatuses.Where(r => r.Recruiter == recruiter).Where(r => r.Student == student);
+            var pstatus = _context.PipelineStatus.Where(r => r.Recruiter == recruiter).Where(r => r.Student == student);
+            var pipelineStatus = new PipelineStatus
+            {
+                Student = student,
+                Recruiter = recruiter,
+                Status = status,
+                Id = Guid.NewGuid().ToString()
+            };
             if (pstatus == null || !pstatus.Any())
             {
-                _context.PipelineStatus.Add(new PipelineStatus(student, recruiter, status));
-                Save();
+                lock (_context)
+                {
+                    _context.PipelineStatus.Add(pipelineStatus);
+                    _context.SaveChanges();
+                }
                 return true;
             }
-            _context.PipelineStatus.Remove(pstatus.First());
-            Save();
-            _context.PipelineStatus.Add(new PipelineStatus(student, recruiter, status));
-            Save();
+            if (pstatus.First() != null)
+            {
+                lock (_context)
+                {
+                    _context.PipelineStatus.Remove(pstatus.First());
+                    _context.SaveChanges();
+                }
+            }
+            lock (_context)
+            {
+                _context.PipelineStatus.Add(pipelineStatus);
+                _context.SaveChanges();
+            }
+            //Save();
             return true;
         }
         override async public Task<bool> SetPipelineStatus(Recruiter recruiter, Student student, string status)
@@ -186,7 +205,7 @@ namespace RecruiterPathway.Repository
             if (!_context.WatchList.Where(w => w.Recruiter == recruiter).Where(s => s.Student == student).Any())
             {
                 _context.WatchList.Add(watch);
-                Save();
+                _context.SaveChanges();
             }
         }
         override async public Task RemoveWatch(Recruiter recruiter, Student student) 
@@ -201,7 +220,7 @@ namespace RecruiterPathway.Repository
                 return;
             }
            _context.WatchList.Remove(recruiter.WatchList.FirstOrDefault(w => w.Student == student));
-           Save();
+           _context.SaveChanges();
         }   
     }
 }
